% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_blocks.R
\name{cl_blocks}
\alias{cl_blocks}
\title{Load the CL dataset and links it with the INE blocks}
\usage{
cl_blocks(path, cve_edo)
}
\arguments{
\item{path}{It is the path where the dataset are hosted}

\item{cve_edo}{is the state id (INEGI)}
}
\value{
a sf object
}
\description{
Load the CL dataset and links it with the INE blocks
}
\details{
xxx
}
\examples{
cl_blocks(path = "/Users/felipedzul/Dropbox/CENAPRECE/2025/14_jalisco", cve_edo = "14")
}
\references{
xxxxx
}
\seealso{
\link[vectoRcontrol]{cl_tb}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
