% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_cl.R
\name{tb_cl}
\alias{tb_cl}
\title{Generate a larval control table}
\usage{
tb_cl(path, jur = NULL, state)
}
\arguments{
\item{path}{is the path of files.}

\item{jur}{is the Jurisdiccion.}

\item{state}{is the logical value, when state = TRUE the output is statal, else jurisdiccional.}
}
\value{
a table.
}
\description{
Generate a larval control table
}
\details{
xxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[formattable]{formattable}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
